fix(angular): add serve static target more intentionally #27854
