+++ packages/angular/src/generators/application/lib/index.ts
@@ -10,3 +10,4 @@ export * from './normalized-schema';
 export * from './set-app-strict-default';
 export * from './set-generator-defaults';
 export * from './update-editor-tsconfig';
+export * from './add-serve-static-target';
